Move python shell related stuff to libyt_python_shell.cpp.
