+++ src/yt_initialize.cpp
@@ -83,8 +83,8 @@ int yt_initialize(int argc, char* argv[], const yt_param_libyt* param_libyt) {
 
     // set python exception hook and set not-yet-done error msg
 #ifdef INTERACTIVE_MODE
-    if (func_status_list::set_exception_hook() == YT_FAIL) return YT_FAIL;
-    if (func_status_list::init_not_done_err_msg() != YT_SUCCESS) return YT_FAIL;
+    if (LibytPythonShell::set_exception_hook() == YT_FAIL) return YT_FAIL;
+    if (LibytPythonShell::init_not_done_err_msg() != YT_SUCCESS) return YT_FAIL;
 #endif
 
     LibytProcessControl::Get().libyt_initialized = true;
